pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "javaxFlash"
version = "3.0.0"
description = "Lightweight Python AI client/router with clean provider architecture, retries, and structured output."
readme = "README.md"
license = "MIT"
authors = [
  { name = "Javas Antony" }
]
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "tavily-python>=0.7.0",
]
keywords = ["ai", "client", "router", "sdk", "gemini", "deepseek"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0",
]

[tool.setuptools]
packages = ["javaxFlash"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[project.urls]
"Homepage" = "https://pypi.org/project/javaxFlash/"
